requirements.txt
# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# LLM APIs
openai>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Statistical validation
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization & reporting
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
tenacity>=8.2.0

